pyproject.toml
[tool.poetry]
name = "aldangold-gendiff"
version = "1.1.0"
description = "Difference calculator"
authors = ["aldangold <meloshnikoff@gmail.com>"]
packages = [
  { include = "gendiff" },
]
repository = "https://github.com/aldangold/python-project-lvl2"
classifiers = [
  "Environment :: Console"
]

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.dependencies]
python = "^3.6"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"
pytest-cov = "^2.10.0"

[[tool.poetry.source]]
name = "ppt"
url = "https://test.pypi.org/legacy/"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

